Address settings lazy load review
